048-YOLO_DeformableDETR_behavior_study/requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.35
timm
numpy
pillow
pycocotools
tqdm
opencv-python
matplotlib
